### 2021-06-16 #3378 @whedon
[PRE REVIEW]: Cutepeaks: A modern viewer for Sanger trace file
**Submitting author:** @dridk (<a href="http://orcid.org/0000-0002-4563-7537">sacha schutz</a>)
**Repository:** <a href="https://github.com/labsquare/CutePeaks" target ="_blank">https://github.com/labsquare/CutePeaks</a>
**Version:** 0.2.3
