### 2021-06-17 #3378 @danielskatz
Thanks @lpantano - just FYI, for submissions with a `waitlisted` tag, you should also remove that tag when you take on a submission.
