### 2021-06-17 #3378 @lpantano
@dridk, could you add or point in the readme to the docs for a quick start steps by step. I see that you have examples in the repository, and I found that reviewer is faster if you add documentation to show how to load an exmaple, and the few features that you can do with your tool. Thanks!
