### 2021-06-17 f09e4fd
Update README.md
Add documentation for JOSS peer review 
https://github.com/openjournals/joss-reviews/issues/3378
