### 2021-06-17 #3378 @dridk
@lpantano Done ! https://github.com/labsquare/CutePeaks/commit/f09e4fd464dce80883e200ecda6847e267f3fca4.
By the way, you have a binary available for each platform. It should be easy to run it :https://labsquare.github.io/CutePeaks/
