### 2021-06-18 #3378 @otre99
I'd love to help you, but I can't at the moment. I'm sorry about that

El jue, 17 jun 2021 a las 10:38, sacha schutz ***@***.***>)
escribió:

> @lpantano <https://github.com/lpantano> Done ! ***@***.***
> <https://github.com/labsquare/CutePeaks/commit/f09e4fd464dce80883e200ecda6847e267f3fca4>
> .
> By the way, you have a binary available for each platform. It should be
> easy to run it :https://labsquare.github.io/CutePeaks/
>
