### 2021-06-28 #3378 @lpantano
@tobiasrausch, would you like to review this tool for JOSS? Read more about the criteria here:  (https://joss.readthedocs.io/en/latest/review_criteria.html) Thanks!
