### 2021-07-08 #3457 @tobiasrausch
proof, line 27: The regular expression `AC.+T` should probably be `AC+T`
