### 2021-07-08 #3457 @wdecoster
Hi @dridk, @tobiasrausch, and @lpantano

This is the first time I review for JOSS, so please let me know if I make a mistake. The process appears to be very clear though.

The installation instructions for Linux mention `chmod +x cutepeaks-0.2.0-linux-x86_64.appimage`, while the name of the binary is just `cutepeaks-x86_64.AppImage`.
Upon calling the executable (`./cutepeaks-x86_64.AppImage`) I get:
```
add font
./cutepeaks-x86_64.AppImage: symbol lookup error: /lib/x86_64-linux-gnu/libfontconfig.so.1: undefined symbol: FT_Done_MM_Var
```

My operating system is Ubuntu 21.04.

Cheers,
Wouter
