### 2021-07-08 #3457 @dridk
@wdecoster sorry for that ! AppImage has been build on a previous system.
Here is a version for your system :
https://github.com/labsquare/CutePeaks/releases/download/0.2.3/cutepeaks-ubuntu_21-04-x86_64.AppImage

Should works ! I just tested it
