### 2021-07-09 37f6d8d
Update README.md
Update documentations ( community and license ) 
https://github.com/openjournals/joss-reviews/issues/3457#issue-938751386
