+++ README.md
@@ -76,3 +76,18 @@ Once the file is open, cutepeaks allows you to :
 - Search for a regular expression pattern. Open the "Find Sequence..." from the edit menu
 - Export trace or sequence to different format. ( e.g: Fasta, CSV, SVG or PNG image ) 
 
+## Contributions / Bugs
+Cutepeaks is a new project and all contributors are welcome
+
+### Issues
+If you found a bug or have a feature request, you can report it from the [Github isse trackers](https://github.com/labsquare/cutepeaks/issues).
+
+### Chat 
+You can join us [on discord](https://discord.gg/7sSH4VSPKK). We are speaking french right now, but we can switch to english. 
+
+## Licenses
+This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+
+This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.txt.
