### 2021-07-20 #3457 @wdecoster
Hi @dridk, looking at the checklist both the paper and documentation miss a "statement of need", and the paper should additionally get a comparison to the "State of the field" - could you look into adding these sections?
