+++ paper/paper.bib
@@ -25,3 +25,26 @@ issn={1471-2164},
 doi={10.1186/s12864-020-6635-8},
 url={https://doi.org/10.1186/s12864-020-6635-8}
 }
+
+@Article{Seqtrace,
+author={Stucky, Brian J.},
+title={SeqTrace: a graphical tool for rapidly processing DNA sequencing chromatograms},
+journal={Journal of biomolecular techniques : JBT},
+year={2012},
+month={Sep},
+publisher={Association of Biomolecular Resource Facilities},
+volume={23},
+number={3},
+pages={90-93},
+keywords={bioinformatics; sequence analysis; software; Base Sequence; Chromatography; *Computer Graphics; Consensus Sequence; Molecular Sequence Annotation; *Sequence Analysis, DNA; *Software},
+abstract={Modern applications of Sanger DNA sequencing often require converting a large number of chromatogram trace files into high-quality DNA sequences for downstream analyses. Relatively few nonproprietary software tools are available to assist with this process. SeqTrace is a new, free, and open-source software application that is designed to automate the entire workflow by facilitating easy batch processing of large numbers of trace files. SeqTrace can identify, align, and compute consensus sequences from matching forward and reverse traces, filter low-quality base calls, and end-trim finished sequences. The software features a graphical interface that includes a full-featured chromatogram viewer and sequence editor. SeqTrace runs on most popular operating systems and is freely available, along with supporting documentation, at http://seqtrace.googlecode.com/.},
+note={22942788[pmid]},
+note={PMC3413935[pmcid]},
+note={12-2303-004[PII]},
+issn={1943-4731},
+doi={10.7171/jbt.12-2303-004},
+url={https://pubmed.ncbi.nlm.nih.gov/22942788},
+url={https://doi.org/10.7171/jbt.12-2303-004},
+language={eng}
+}
+
