### 2021-07-20 5f1355e
Update paper.md
shortcut the paper
