+++ paper/paper.md
@@ -26,36 +26,30 @@ bibliography: paper.bib
 ---
 
 # Summary
-In the present note, we introduce an original Sanger trace viewer steered by a modern and user-friendly graphical interface. Unlike other software, CutePeaks comes with two new features: searching for a regular expression and exporting the traces to svg.    
-CutePeaks is available as a standalone application for Linux, MacOS and Windows at    
-[https://labsquare.github.io/CutePeaks/](https://labsquare.github.io/CutePeaks/).
+Cutepeaks is a standalone Sanger trace viewer steered by a modern and user-friendly UI. Unlike other software, CutePeaks comes with two new features: searching for a regular expression and exporting the traces to svg.    
+CutePeaks is available for Linux, MacOS and Windows at [https://labsquare.github.io/CutePeaks/](https://labsquare.github.io/CutePeaks/).
 
 # Statement of need
 Despite the major use of Next Generation Sequencing, the Sanger method is still widely used in genetic labs as the gold standard to read target DNA sequences. Very few opensource software is available to explore Sanger trace data and most of labs staff still rely on proprietary software. Moreover, they are not always user-friendly and lack modern look and feel. 
 
 # State of fields
-4peaks[@4Peaks] is a software widely used by biologists that benefits from a user-friendly and modern user interface. Sadly, it is only available on MacOS and source code is not opened to community enhancement. Seqtrace[@seqtrace] is the only standalone application published with its source code. However, it is written with GTK framework in Python 2, the latter being deprecated and slower than C++. 
+4peaks[@4Peaks] is a software widely used by biologists that benefits from a nice User interface. Sadly, it is only available on MacOS and source code is not opened to community enhancement. Seqtrace[@seqtrace] is the only standalone and opensource application we could find. However, it is written with GTK framework in Python 2, the latter being deprecated and slower than C++. 
 
 # Software overview
 ![CutePeaks screenshot with regular expression search bar.\label{fig:example}](figure.png)
 
-CutePeaks is a modern standalone cross-platform GUI (Graphical User Interface) application implemented in C++ using the open source Qt5 framework. It can read FSA and ABIF file formats, and display the chromatogram with standard controllers.
+CutePeaks is a cross-platform application implemented in C++ using the open source Qt5 framework. It can read FSA and ABIF file formats, and display the chromatogram with standard controllers.
 The chromatogram is displayed in an interactive window allowing the user to move along the trace. It can also re-scale dynamically the plot using two slider controllers. Finger gestures are also available for scrolling upon using a touch screen.
 Similarily to 4peaks software[@4Peaks], Phred quality scores are displayed behind the trace as a blue histogram. Base calling is displayed at the top of the viewing window, along with adjustable amino-acid translation.
-The trace can be used as with a text editor using standard keyboard shortcuts. The user can select an area and copy it to the clipboard, as well as cutting parts of the trace. It is also possible to revert the full trace by applying a revert/complement transformation.
-An original feature of CutePeaks is the possibility to search for a sequence in the trace using a regular expression. This is especially useful to search for a sequence pattern. For example, the query `A[CG]T` will search for all instances of `ACT` or `AGT`. The query `AC.+T` will select all instances of the form `ACT`, `ACCT`, `ACCCCT` etc.
-Finally, the trace data can be exported to different formats, such as FASTA or SVG image, the latter being particularly useful for resolution-independent illustration.
+The trace can be used as with a text editor. Navigating along the trace, copying the sequence to the clipboard or cutting it is done using standard keyboard shortcuts. Revert/complement is also possible.
[…]
