+++ paper/paper.bib
@@ -26,7 +26,7 @@ doi={10.1186/s12864-020-6635-8},
 url={https://doi.org/10.1186/s12864-020-6635-8}
 }
 
-@Article{Seqtrace,
+@Article{seqtrace,
 author={Stucky, Brian J.},
 title={SeqTrace: a graphical tool for rapidly processing DNA sequencing chromatograms},
 journal={Journal of biomolecular techniques : JBT},
