### 2021-07-20 #3457 @dridk
Hi @wdecoster ,
I reformulated a bit the paper to have the required section.

Hi @tobiasrausch
I added the contributions section here :
https://github.com/labsquare/CutePeaks#contributions--bugs

For testing, I don't know if it is enough, but I compile the code each time from github action.
