### 2021-07-21 #3457 @tobiasrausch
Hi @dridk, I think GitHub defaults to a contributing.md file in case you want to add this. Here is the [link for your repo](https://github.com/labsquare/CutePeaks/community). Did you have a chance to check your regex in the paper? I believe this should be AC+T and not AC.+T from the examples you show afterwards.
