### 2021-07-21 9f59f02
Fix regexp
https://github.com/openjournals/joss-reviews/issues/3457#issuecomment-884091094
