### 2021-07-21 3a4c9f8
Create contribution.md
See https://github.com/openjournals/joss-reviews/issues/3457#issuecomment-884091094
