+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# Contributing
+All assistance is welcome. You can contribute to the project from the following topic:
+
+# Issues
+Bug or feature requests can report from [github issue tracker](https://github.com/labsquare/CutePeaks).
+
+# Pull request
+Fell free to create pull request. Compilation test from [github](https://github.com/labsquare/CutePeaks/actions/workflows/c-cpp.yml) action are triggered from any pull request.
+
+# Coding style
+Please respect the Qt coding style described [here](https://wiki.qt.io/Qt_Coding_Style)
+
+# Translation
+Cutepeaks is only translated in French. You can translate to your language by using [Qt linguist](https://doc.qt.io/qt-5/linguist-manager.html)
+
+# Chat
+You can join us [on discord](https://discord.com/invite/7sSH4VSPKK). We are speaking french right now, but we can switch to english.
