### 2021-07-22 #3457 @wdecoster
I think for a tool like this the best automated testing is testing if it compiles - so I think that requirement is fulfilled.
