### 2021-07-28 ee3e919
Add author and fix typos
