+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Cutepeaks: A modern viewer for Sanger trace file'
+title: 'CutePeaks: A modern viewer for Sanger trace file'
 tags:
   - Python
   - Sanger
@@ -13,12 +13,19 @@ authors:
   - name: Charles Monod-Broca^[co-first author]
     orcid: 0000-0003-4095-8099
     affiliation: "2"
+  - name: Anne-Sophie Denommé-Pichon
+    orcid: 0000-0002-8986-8222
+    affiliation: "3,4"
 
 affiliations:
  - name: CHRU Brest 29200, Brest, France
    index: 1
  - name: Univ Brest, Inserm, EFS, UMR 1078, GGB, 29200
    index: 2
+ - name: Unité Fonctionnelle Innovation en Diagnostic génomique des maladies rares, FHU-TRANSLAD, CHU Dijon Bourgogne, Dijon, France
+   index: 3
+ - name: UMR1231 GAD, Inserm - Université Bourgogne-Franche Comté, Dijon, France
+   index: 4
 
 date: 15 June 2021
 bibliography: paper.bib
@@ -26,33 +33,33 @@ bibliography: paper.bib
 ---
 
 # Summary
-Cutepeaks is a standalone Sanger trace viewer steered by a modern and user-friendly UI. Unlike other software, CutePeaks comes with two new features: searching for a regular expression and exporting the traces to svg.    
-CutePeaks is available for Linux, MacOS and Windows at [https://labsquare.github.io/CutePeaks/](https://labsquare.github.io/CutePeaks/).
+CutePeaks is a standalone Sanger trace viewer steered by a modern and user-friendly UI. Unlike other software, CutePeaks comes with two new features: searching for a regular expression and exporting the traces to SVG.    
+CutePeaks is available for Linux, macOS and Windows at [https://labsquare.github.io/CutePeaks/](https://labsquare.github.io/CutePeaks/).
 
 # Statement of need
-Despite the major use of Next Generation Sequencing, the Sanger method is still widely used in genetic labs as the gold standard to read target DNA sequences. Very few opensource software is available to explore Sanger trace data and most of labs staff still rely on proprietary software. Moreover, they are not always user-friendly and lack modern look and feel. 
+Despite the major use of next generation sequencing, the Sanger method is still widely used in genetic labs as the gold standard to read target DNA sequences. Very few open source software is available to explore Sanger trace data and most of labs staff still rely on proprietary software. Moreover, they are not always user-friendly and lack modern look and feel. 
 
 # State of fields
-4peaks[@4Peaks] is a software widely used by biologists that benefits from a nice User interface. Sadly, it is only available on MacOS and source code is not opened to community enhancement. Seqtrace[@seqtrace] is the only standalone and opensource application we could find. However, it is written with GTK framework in Python 2, the latter being deprecated and slower than C++. 
+4peaks[@4Peaks] is software widely used by biologists that benefits from a nice user interface. Sadly, it is only available on macOS and source code is not open to the community enhancement. Seqtrace[@seqtrace] is the only standalone and open source application we could find. However, it is written with the GTK framework in Python 2, the latter being deprecated and slower than C++. 
 
 # Software overview
 ![CutePeaks screenshot with regular expression search bar.\label{fig:example}](figure.png)
 
 CutePeaks is a cross-platform application implemented in C++ using the open source Qt5 framework. It can read FSA and ABIF file formats, and display the chromatogram with standard controllers.
-The chromatogram is displayed in an interactive window allowing the user to move along the trace. It can also re-scale dynamically the plot using two slider controllers. Finger gestures are also available for scrolling upon using a touch screen.
[…]
