### 2021-07-29 912cf4b
Update CITATION.cff
Update citation
