+++ CITATION.cff
@@ -16,5 +16,5 @@ authors:
     
 title: labsquare/CutePeaks
 version: 0.2.3
-date-released: 2021-07-29
+date-released: 2021-07-28
             
