+++ README.md
@@ -1,6 +1,9 @@
 # CutePeaks
 
 [![C/C++ CI](https://github.com/labsquare/CutePeaks/actions/workflows/c-cpp.yml/badge.svg)](https://github.com/labsquare/CutePeaks/actions/workflows/c-cpp.yml)
+[![DOI](https://zenodo.org/badge/79352493.svg)](https://zenodo.org/badge/latestdoi/79352493)
+
+
 
 https://labsquare.github.io/CutePeaks/
 
