### 2021-07-29 #3457 @dridk
- Tag Release:
https://github.com/labsquare/CutePeaks/releases/tag/0.2.3
- Zenodo :
https://zenodo.org/record/5145153#.YQJ0i346_RY
- DOI
10.5281/zenodo.5145153
