+++ CITATION.cff
@@ -14,7 +14,7 @@ authors:
     affiliation: UMR1231 GAD, Inserm - Université Bourgogne-Franche Comté, Dijon, France
     orcid: https://orcid.org/0000-0002-8986-8222
     
-title: CutePeaks: A modern viewer for Sanger trace file
+title: 'CutePeaks: A modern viewer for Sanger trace file'
 version: 0.2.3
 date-released: 2021-07-28
             
