### 2021-08-06 #3457 @dridk
Thanks you all !
I was really impressed with your reviewing system which is very smart ! The markdown paper, the github issue system and your awesome whedon ! It would be nice if all scientific journals did the same !
